:sparkles: Make get and first cache keys match
Previously, the cache key generated for get() and first() were different
because the names of the dictionary key containing the parms to bind
where different ('id' if get, 'param' if first).

This change injects the bound params into the query and then hashes
that.

The purpose for this change, was that previously, it was impossible to
invalidate cache for a get() or get_or_404().